hotfix: refactored classes to follow SRP.
