+++ src/test/java/com/seyed/ali/timeentryservice/TimeEntryQueryServiceApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TimeEntryServiceApplicationTests {
+class TimeEntryQueryServiceApplicationTests {
 
     @Test
     void contextLoads() {
